Aspectj implementation in test scope
